Fix warnings and ensure type safety in benchmark tests
- Suppressed warning C4996 in macro and micro benchmark tests when reading the POLYBENCH_MODE environment variable.
- Added static_cast<int> to ensure type safety when accumulating values in the threading services test.

+++ tests/unit/runtime/threading_services_test.cpp
@@ -150,7 +150,7 @@ TEST_CASE("Threading - Work Stealing Scheduler", "[threading][worksteal]") {
         std::atomic<int> sum{0};
         
         scheduler.ParallelFor(0, 1000, [&sum](size_t i) {
-            sum += i;
+            sum += static_cast<int>(i);
         });
         
         REQUIRE(sum == 499500);  // 0 + 1 + ... + 999
